Removes publishapplicationdetermination, renames to requestevidencedetermination
